物料库存及物料操作完成

+++ src/components/UploadImage.vue
@@ -15,11 +15,15 @@ import { ref, watch } from 'vue'
 import { getStsAPI } from "../apis/util";
 import type { UploadProps, UploadUserFile } from 'element-plus'
 import Snowflake from "@axihe/snowflake"
+import { ElMessage } from 'element-plus'
+
 const emit = defineEmits(["uploadImage"])
 const props = defineProps({
     dialog: { type: Boolean },
-    confirmImage: { type: Boolean }
+    confirmImage: { type: Boolean },
+    uploaded: { type: String, default: null }
 });
+const original = ref()
 
 //雪花算法生成id
 //#region 
@@ -38,6 +42,7 @@ const idWorker = new Snowflake(config.worker_id, config.datacenter_id);
 
 const fileList = ref<UploadUserFile[]>([])
 
+
 const getSts = async () => {
     const res = await getStsAPI()
 
@@ -52,6 +57,7 @@ const getSts = async () => {
 getSts()
 
 const uidToFileNameMap = ref({});
+const fileToDelete = ref<UploadUserFile[]>([])
 
 const customHttpRequest = async (options) => {
     const { file, data, headers, withCredentials } = options;
@@ -90,26 +96,44 @@ const customHttpRequest = async (options) => {
 
 const handleRemove: UploadProps['onRemove'] = async (uploadFile, uploadFiles) => {
     const fileName = uidToFileNameMap.value[uploadFile.uid];
-    const s3 = new AWS.S3({
-        apiVersion: '2006-03-01',
-        endpoint: "https://eos-chengdu-1.cmecloud.cn",
-        signatureVersion: 'v2',
-        sslEnabled: false,
-    });
-    const params = {
-        Bucket: 'cmc/receipt',
-        Key: fileName,
-    };
-    try {
-        await s3.deleteObject(params).promise();
-        console.log('File delete successfully');
+
+
+    if (original.value.hasOwnProperty(uploadFile.uid)) {
+        console.log("delete original files");
+        fileToDelete.value.push(uploadFile);
         delete uidToFileNameMap.value[uploadFile.uid];
+        console.log(fileToDelete);
 
-    } catch (error) {
-        console.error('Error deleting file:', error);
     }
-    // console.log(uploadFile, uploadFiles)
+    else {
+        const s3 = new AWS.S3({
+            apiVersion: '2006-03-01',
+            endpoint: "https://eos-chengdu-1.cmecloud.cn",
+            signatureVersion: 'v2',
+            sslEnabled: false,
+        });
+        const params = {
+            Bucket: 'cmc/receipt',
+            Key: fileName,
+        };
+        try {
+            await s3.deleteObject(params).promise();
+            console.log('File delete successfully');
+            ElMessage({
+                type: 'success',
+                message: '删除成功',
+            })
+            delete uidToFileNameMap.value[uploadFile.uid];
 
+        } catch (error) {
+            ElMessage({
+                type: 'info',
+                message: '删除取消',
+            })
+            console.error('Error deleting file:', error);
+        }
+        // console.log(uploadFile, uploadFiles)
+    }
 }
 const dialogImageUrl = ref('')
 const dialogVisible = ref(false)
@@ -125,9 +149,12 @@ watch([() => props.dialog, () => props.confirmImage], async ([dialogValue, confi
     console.log('dialog value changed:', dialogValue);
     console.log('confirmImage value changed:', confirmImageValue);
 
+    //如果弹框关闭且用户未点击确认
     if (!dialogValue && !confirmImageValue) {
         // 执行删除文件操作或其他操作
         //   Iterate through uidToFileNameMap and delete files from AWS S3
+        console.log("dialog false and confirmImage false");
+
         const s3 = new AWS.S3({
             apiVersion: '2006-03-01',
             endpoint: "https://eos-chengdu-1.cmecloud.cn",
@@ -136,7 +163,39 @@ watch([() => props.dialog, () => props.confirmImage], async ([dialogValue, confi
         });
 
         for (const uid in uidToFileNameMap.value) {
-            const fileName = uidToFileNameMap.value[uid];
+            //不删除开启弹框前已有的图片
+            if (!original.value.hasOwnProperty(uid)) {
+                const fileName = uidToFileNameMap.value[uid];
+                const params = {
+                    Bucket: 'cmc/receipt',
+                    Key: fileName,
+                };
+
+                try {
+                    await s3.deleteObject(params).promise();
+                    console.log(`File ${fileName} deleted successfully`);
+                    delete uidToFileNameMap.value[uid];
+
+                } catch (error) {
+                    console.error(`Error deleting file ${fileName}:`, error);
+                }
+            }
+        }
+    }
+    //如果弹框关闭，用户提交确认按钮，删除要删除的图片
+    if (!dialogValue && confirmImageValue) {
+        // Execute delete file operation for files in fileToDelete
+        const s3 = new AWS.S3({
+            apiVersion: '2006-03-01',
+            endpoint: "https://eos-chengdu-1.cmecloud.cn",
+            signatureVersion: 'v2',
+            sslEnabled: false,
+        });
+
+        fileToDelete.value.forEach(async (fileToDeleteItem) => {
+            const fileName = fileToDeleteItem.name
+            console.log(fileName);
+
             const params = {
                 Bucket: 'cmc/receipt',
                 Key: fileName,
@@ -145,25 +204,55 @@ watch([() => props.dialog, () => props.confirmImage], async ([dialogValue, confi
             try {
                 await s3.deleteObject(params).promise();
                 console.log(`File ${fileName} deleted successfully`);
-                delete uidToFileNameMap.value[uid];
-
+                // delete uidToFileNameMap.value[fileToDeleteItem.uid];
             } catch (error) {
                 console.error(`Error deleting file ${fileName}:`, error);
             }
-        }
+        });
+
+        // Clear the fileToDelete list
+        fileToDelete.value = [];
+    }
+    // 其他操作...
+});
+
+// console.log("original upload", props.uploaded);
+
+watch(() => props.uploaded, (newval, oldval) => {
+    // console.log("props uploaded changed");
+    // console.log("new", newval);
+    // console.log("old", oldval);
+
+    if (newval != null && oldval == null) {
+        // console.log("first time");
+        original.value = JSON.parse(newval)
+        uidToFileNameMap.value = JSON.parse(newval)
+
+
+        // console.log(uidToFileNameMap);
+        fileList.value = Object.keys(uidToFileNameMap.value).map(uid => ({
+            uid: uid,
+            name: uidToFileNameMap.value[uid],
+            status: "success",
+            url: "https://cmc.eos-chengdu-1.cmecloud.cn/receipt/" + uidToFileNameMap.value[uid] // Replace with the actual URL
+        }));
+
+        // console.log(fileList);
+
     }
 
     // 其他操作...
-});
+}, { deep: true });
 
 watch(
     uidToFileNameMap,
     (newMap, oldMap) => {
         // console.log('uidToFileNameMap changed:');
-        // console.log('New uidToFileNameMap:', newMap);
+        console.log('New uidToFileNameMap:', newMap);
+        // console.log('Old uidToFileNameMap:', oldMap);
 
-        // You can perform actions here based on the changes in the uidToFileNameMap
         emit("uploadImage", newMap);
+
 
     },
     { deep: true } // Enable deep monitoring
